// src/gql/graphql.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any }
}
// ...
export type CreateAccountOutput = {
  __typename?: 'CreateAccountOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}
// ...
export type CreateOrderOutput = {
  __typename?: 'CreateOrderOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}

export type CreatePaymentInput = {
  restaurantId: Scalars['Int']['input']
  transactionId: Scalars['String']['input']
}

export type CreatePaymentOutput = {
  __typename?: 'CreatePaymentOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}

export type CreateRestaurantInput = {
  address: Scalars['String']['input']
  categoryName: Scalars['String']['input']
  coverImage: Scalars['String']['input']
  name: Scalars['String']['input']
}

export type CreateRestaurantOutput = {
  __typename?: 'CreateRestaurantOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}

export type DeleteDishInput = {
  dishId: Scalars['Int']['input']
}

export type DeleteDishOutput = {
  __typename?: 'DeleteDishOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}

export type DeleteRestaurantInput = {
  restaurantId: Scalars['Float']['input']
}

export type DeleteRestaurantOutput = {
  __typename?: 'DeleteRestaurantOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}
// ...
export type EditDishOutput = {
  __typename?: 'EditDishOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}

export type EditOrderInput = {
  id: Scalars['Float']['input']
  status: OrderStatus
}

export type EditOrderOutput = {
  __typename?: 'EditOrderOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}

export type EditProfileInput = {
  email?: InputMaybe<Scalars['String']['input']>
  password?: InputMaybe<Scalars['String']['input']>
}

export type EditProfileOutput = {
  __typename?: 'EditProfileOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
  user: User
}

export type EditRestaurantInput = {
  address?: InputMaybe<Scalars['String']['input']>
  categoryName?: InputMaybe<Scalars['String']['input']>
  coverImage?: InputMaybe<Scalars['String']['input']>
  name?: InputMaybe<Scalars['String']['input']>
  restaurantId: Scalars['Float']['input']
}

export type EditRestaurantOutput = {
  __typename?: 'EditRestaurantOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}
// ...
export type LoginInput = {
  email: Scalars['String']['input']
  password: Scalars['String']['input']
}

export type LoginOutput = {
  __typename?: 'LoginOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
  token?: Maybe<Scalars['String']['output']>
}

export type Mutation = {
  __typename?: 'Mutation'
  createAccount: CreateAccountOutput
  createDish: CreateRestaurantOutput
  createOrder: CreateOrderOutput
  createPayment: CreatePaymentOutput
  createRestaurant: CreateRestaurantOutput
  deleteDish: DeleteDishOutput
  deleteRestaurant: DeleteRestaurantOutput
  editDish: EditDishOutput
  editOrder: EditOrderOutput
  editProfile: EditProfileOutput
  editRestaurant: EditRestaurantOutput
  login: LoginOutput
  takeOrder: TakeOrderOutput
  verifyEmail: VerifyEmailOutput
}
// ...
export type TakeOrderOutput = {
  __typename?: 'TakeOrderOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}
// ...
export type VerifyEmailOutput = {
  __typename?: 'VerifyEmailOutput'
  error?: Maybe<Scalars['String']['output']>
  ok: Scalars['Boolean']['output']
}

export type LoginMutationVariables = Exact<{
  loginInput: LoginInput
}>

export type LoginMutation = {
  __typename?: 'Mutation'
  login: {
    __typename?: 'LoginOutput'
    ok: boolean
    token?: string | null
    error?: string | null
  }
}